refactor: rename 'server' dir into 'services'
